+++ tests/codec/test_markdown.py
@@ -4,7 +4,9 @@
 from __future__ import annotations
 
 from kif_lib import (
+    Context,
     DeprecatedRank,
+    ExternalId,
     ExternalIdDatatype,
     Filter,
     IRI,
@@ -12,6 +14,7 @@ from kif_lib import (
     Item,
     ItemDatatype,
     KIF_Object,
+    Lexeme,
     LexemeDatatype,
     NormalRank,
     NoValueSnak,
@@ -85,8 +88,7 @@ class Test(TestCase):
             wd.Q('_'), self.md_sexp('Item', self.md_link(
                 'wd:Q_', wd.Q('_').iri.content)))
         # unknown prefix
-        self.assert_to_markdown(
-            Item('x'), self.md_sexp('Item', self.md_link('x', 'x')))
+        self.assert_to_markdown(Item('x'), self.md_sexp('Item', 'x'))
 
     def test_property_to_markdown(self) -> None:
         # known prefix: know label
@@ -99,24 +101,41 @@ class Test(TestCase):
                 'wd:P_', wd.P('_').iri.content)))
         # unknown prefix
         self.assert_to_markdown(
-            Property('x'), self.md_sexp('Property', self.md_link('x', 'x')))
+            Property('x'), self.md_sexp('Property', 'x'))
+
+    def test_lexeme_to_markdown(self) -> None:
+        # known prefix: know lemma, category, language
+        with Context():
+            wd.English.register(label='English')
+            wd.verb.register(label='verb')
+            wd.L(33).register(
+                lemma='change', category=wd.verb, language=wd.English)
+            self.assert_to_markdown(
+                wd.L(33), self.md_sexp(
+                    'Lexeme',
+                    self.md_link('change', wd.L(33).iri.content) + ' ('
+                    + self.md_link('English', wd.English.iri.content) + ' '
+                    + self.md_link("verb", wd.verb.iri.content) + ')'))
+        # known prefix: unknown lemma, category, language
+        with Context():
+            self.assert_to_markdown(
+                wd.L(33), self.md_sexp('Lexeme', self.md_link(
+                    'wd:L33', wd.L(33).iri.content)))
+        # unknown prefix
+        self.assert_to_markdown(
+            Lexeme('x'), self.md_sexp('Lexeme', 'x'))
 
     def test_iri_to_markdown(self) -> None:
         # known prefix
         self.assert_to_markdown(
             IRI('http://www.wikidata.org/entity/Q5'),
-            self.md_link('wd:Q5', 'http://www.wikidata.org/entity/Q5'))
+            'http://www.wikidata.org/entity/Q5')
         # unknown prefix: absolute iri
         self.assert_to_markdown(
-            IRI('http://prefix/x'),
-            self.md_link('http://prefix/x'))
+            IRI('http://prefix/x'), 'http://prefix/x')
         # unknown prefix: relative iri
-        self.assert_to_markdown(
-            IRI('prefix/x'),
-            self.md_link('prefix/x', 'prefix/x'))
-        self.assert_to_markdown(
-            IRI('x'),
-            self.md_link('x', 'x'))
+        self.assert_to_markdown(IRI('prefix/x'), 'prefix/x')
+        self.assert_to_markdown(IRI('x'), 'x')
 
     def test_text_to_markdown(self) -> None:
         self.assert_to_markdown(Text('abc'), '"abc"@en')
@@ -132,81 +151,100 @@ class Test(TestCase):
         self.assert_to_markdown(String('`'), r'"\`"')
         self.assert_to_markdown(String('[]'), r'"\[]"')
 
+    def test_external_id_to_markdown(self) -> None:
+        self.assert_to_markdown(ExternalId('abc'), '"abc"')
+        # escape
+        self.assert_to_markdown(ExternalId('_'), r'"\_"')
+        self.assert_to_markdown(ExternalId('*'), r'"\*"')
+        self.assert_to_markdown(ExternalId('`'), r'"\`"')
+        self.assert_to_markdown(ExternalId('[]'), r'"\[]"')
+
     def test_quantity_to_markdown(self) -> None:
+        dalton = self.md_link('dalton', wd.dalton.iri.content)
+        kilogram = self.md_link('kilogram', wd.kilogram.iri.content)
         # amount
-        self.assert_to_markdown(
-            Quantity(0),
-            self.md_sexp('Quantity', 0))
+        self.assert_to_markdown(Quantity(0), '0')
         # amount, unit
         self.assert_to_markdown(
-            Quantity(0, wd.dalton),
-            self.md_sexp('Quantity', 0, wd.dalton))
+            Quantity(0, wd.dalton), '0 ' + dalton)
         # amount, unit, lower bound
         self.assert_to_markdown(
-            Quantity(0, wd.kilogram, -1),
-            self.md_sexp('Quantity', 0, '[-1,∞]', wd.kilogram))
+            Quantity(0, wd.kilogram, -1), '0 [-1,+∞] ' + kilogram)
         # amount, unit, upper bound
         self.assert_to_markdown(
-            Quantity(0, wd.kilogram, None, 1),
-            self.md_sexp('Quantity', 0, '[-∞,1]', wd.kilogram))
+            Quantity(0, wd.kilogram, None, 1), '0 [-∞,1] ' + kilogram)
         # amount, unit, lower bound, upper bound
         self.assert_to_markdown(
-            Quantity(0, wd.kilogram, -1, 1),
-            self.md_sexp('Quantity', 0, '±1', wd.kilogram))
+            Quantity(0, wd.kilogram, -1, 1), '0 ±1 ' + kilogram)
         # amount, lower bound
         self.assert_to_markdown(
-            Quantity(-1, None, 0),
-            self.md_sexp('Quantity', -1, '[0,∞]'))
+            Quantity(-1, None, 0), '-1 [0,+∞]')
         # amount, upper bound
         self.assert_to_markdown(
-            Quantity(-1, None, None, 0),
-            self.md_sexp('Quantity', -1, '[-∞,0]'))
+            Quantity(-1, None, None, 0), '-1 [-∞,0]')
         # amount, lower bound, upper bound
         self.assert_to_markdown(
-            Quantity(-1, None, -1, 0),
-            self.md_sexp('Quantity', -1, '[-1,0]'))
+            Quantity(-1, None, -1, 0), '-1 [-1,0]')
         self.assert_to_markdown(
-            Quantity('2', None, '1', '3'),
-            self.md_sexp('Quantity', 2, '±1'))
+            Quantity('2', None, '1', '3'), '2 ±1')
         self.assert_to_markdown(
-            Quantity('-2', None, '-3', '-1'),
-            self.md_sexp('Quantity', -2, '±1'))
+            Quantity('-2', None, '-3', '-1'), '-2 ±1')
 
     def test_time_to_markdown(self) -> None:
-        # time
         self.assert_to_markdown(
-            Time('2023-09-28'),
-            self.md_sexp('Time', '2023-09-28'))
+            Time('2023-09-28'), '2023-09-28')
         self.assert_to_markdown(
-            Time('2023-09-28T00'),
-            self.md_sexp('Time', '2023-09-28'))
+            Time('2023-09-28T00'), '2023-09-28')
         self.assert_to_markdown(
-            Time('2023-09-28T00:00'),
-            self.md_sexp('Time', '2023-09-28'))
+            Time('2023-09-28T00:00'), '2023-09-28')
         self.assert_to_markdown(
-            Time('2023-09-28T00:00:00'),
-            self.md_sexp('Time', '2023-09-28'))
+            Time('2023-09-28T00:00:00'), '2023-09-28')
         self.assert_to_markdown(
-            Time('2023-09-28T00:00:01'),
-            self.md_sexp('Time', '2023-09-28'))
+            Time('2023-09-28T00:00:01'), '2023-09-28')
         # time, precision
         self.assert_to_markdown(
-            Time('2023-09-28', Time.Precision.DAY),
-            self.md_sexp('Time', '2023-09-28'))
+            Time('2023-09-28', Time.Precision.BILLION_YEARS),
+            '2023-09-28')
         self.assert_to_markdown(
-            Time('2023-09-28', Time.Precision.HOUR),
-            self.md_sexp('Time', '2023-09-28T00:00:00+00:00'))
+            Time('2023-09-28', Time.Precision.HUNDRED_THOUSAND_YEARS),
+            '2023-09-28')
         self.assert_to_markdown(
-            Time('2023-09-28T01', Time.Precision.HOUR),
-            self.md_sexp('Time', '2023-09-28T01:00:00+00:00'))
+            Time('2023-09-28', Time.Precision.TEN_MILLION_YEARS),
+            '2023-09-28')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.MILLION_YEARS),
+            '2023-09-28')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.HUNDRED_THOUSAND_YEARS),
+            '2023-09-28')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.TEN_THOUSAND_YEARS),
+            '2023-09-28')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.MILLENNIA), '3. millennium')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.CENTURY), '21. century')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.YEAR), '2023')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.MONTH), 'September 2023')
+        self.assert_to_markdown(
+            Time('2023-09-28', Time.Precision.DAY), '28 September 2023')
+        self.assert_to_markdown(
+            Time('2023-09-28T01:01', Time.Precision.HOUR),
+            '28 September 2023 at 01:01')
+        self.assert_to_markdown(
+            Time('2023-09-28T00:04', Time.Precision.MINUTE),
+            '28 September 2023 at 00:04')
+        self.assert_to_markdown(
+            Time('2023-09-28T01:02:03', Time.Precision.SECOND),
+            '28 September 2023 at 01:02:03')
         # time, timezone
-        self.assert_to_markdown(
-            Time('2023-09-28', None, 3),
-            self.md_sexp('Time', '2023-09-28'))
+        self.assert_to_markdown(Time('2023-09-28', None, 3), '2023-09-28')
         # time, calendar
         self.assert_to_markdown(
-            Time('2023-09-28', None, None, wd.proleptic_Gregorian_calendar),
-            self.md_sexp('Time', '2023-09-28'))
+            Time('2023-09-28', None, None,
+                 wd.proleptic_Gregorian_calendar), '2023-09-28')
 
     def test_snak_to_markdown(self) -> None:
         self.assert_to_markdown(
@@ -246,7 +284,7 @@ class Test(TestCase):
         self.assert_to_markdown(ref, '''\
 (**ReferenceRecord**
 - (**ValueSnak** (**Property** [country](http://www.wikidata.org/entity/P17)) (**Item** [Brazil](http://www.wikidata.org/entity/Q155)))
-- (**ValueSnak** (**Property** [mass](http://www.wikidata.org/entity/P2067)) (**Quantity** 0))
+- (**ValueSnak** (**Property** [mass](http://www.wikidata.org/entity/P2067)) 0)
 - (**ValueSnak** (**Property** [canonical SMILES](http://www.wikidata.org/entity/P233)) "ABC"))''')  # noqa: E501
 
     def test_filter_to_markdown(self) -> None: